Snap.Internal.Exceptions

Description

An internal Snap module containing the exception that escapes HTTP types.

N.B. this is an internal interface, please don't write user code that depends on it. Interfaces subject to change etc etc etc.

Documentation

data UncatchableException

An exception hierarchy for exceptions that cannot be caught by user-defined error handlers

Constructors

forall e . Exception e => UncatchableException e

uncatchableExceptionToException :: Exception e => e -> SomeException

uncatchableExceptionFromException :: Exception e => SomeException -> Maybe e

data ConnectionTerminatedException

Constructors

ConnectionTerminatedException SomeException

data EscapeHttpException

This exception is thrown if the handler chooses to escape regular HTTP traffic.

Constructors

EscapeHttpException EscapeHttpHandler

type EscapeHttpHandler

Arguments

= ((Int -> Int) -> IO ())
timeout modifier

-> Iteratee ByteString IO ()
socket write end

-> Iteratee ByteString IO ()