Snap.Internal.Debug

Description

An internal Snap module for (optionally) printing debugging messages. Normally debug does nothing, but if you set DEBUG=1 in the environment you'll get debugging messages. We use unsafePerformIO to make sure that the call to getEnv is only made once.

N.B. this is an internal interface, please don't write external code that depends on it.

Documentation

debug :: MonadIO m => String -> m ()

debugErrno :: MonadIO m => String -> m ()

debugIgnore :: MonadIO m => String -> m ()

debugErrnoIgnore :: MonadIO m => String -> m ()